LPS-189386 Consolidate to one class
